code/requirements.txt
numpy
nltk
tensorflow
flask-cors
flask
psycopg2-binary
flask-sqlalchemy
postgres 